+++ src/Polars/Internal/Raw.hs
@@ -72,6 +72,7 @@ module Polars.Internal.Raw
     , phs_expr_sort_by
     , phs_expr_string_function
     , phs_expr_string_function_i64
+    , phs_expr_string_nary_function
     , phs_expr_list_function
     , phs_expr_temporal_function
     , phs_expr_temporal_time_unit
@@ -419,6 +420,9 @@ foreign import ccall unsafe "phs_expr_string_function"
 
 foreign import ccall unsafe "phs_expr_string_function_i64"
     phs_expr_string_function_i64 :: CInt -> CLLong -> Ptr RawExpr -> Ptr (Ptr RawExpr) -> CSize -> Ptr (Ptr RawExpr) -> Ptr (Ptr RawError) -> IO CInt
+
+foreign import ccall unsafe "phs_expr_string_nary_function"
+    phs_expr_string_nary_function :: CInt -> CString -> CBool -> Ptr (Ptr RawExpr) -> CSize -> Ptr (Ptr RawExpr) -> Ptr (Ptr RawError) -> IO CInt
 
 foreign import ccall unsafe "phs_expr_list_function"
     phs_expr_list_function :: CInt -> Ptr RawExpr -> Ptr (Ptr RawExpr) -> CSize -> Ptr (Ptr RawExpr) -> Ptr (Ptr RawError) -> IO CInt